% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{accClimateVariables}
\alias{accClimateVariables}
\title{Describes all the variables available in the database.}
\usage{
accClimateVariables()
}
\value{
A data frame descriptive of the climate variables available in the database.
}
\description{
Describes all the variables available in the database.
}
\examples{
accClimateVariables()
}
