% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{accContinentNames}
\alias{accContinentNames}
\title{Return the list of the continents and associated countries.}
\usage{
accContinentNames()
}
\value{
A list with elements that correspond to the country names of each continent.
}
\description{
Return the list of the continents and associated countries.
}
\examples{
accContinentNames()
}
