% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{accCountryNames}
\alias{accCountryNames}
\title{Return the list of the continents and associated countries.}
\usage{
accCountryNames(continent = NA, dbname = "gbif4crest_02")
}
\arguments{
\item{continent}{A name of continent. Default is NA and returns a list of all
the country names sorted by continent.}

\item{dbname}{The name of the database. Default is gbif4crest_02.}
}
\value{
A list where each element is a vector of corresponding country names.
}
\description{
Return the list of the continents and associated countries.
}
\examples{
accCountryNames()
}
