% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit_pdfsp.R
\name{calib_clim_space}
\alias{calib_clim_space}
\title{Calibrate the distribution of the modern climate space.}
\usage{
calib_clim_space(climate, bin_width)
}
\arguments{
\item{climate}{All the climate values observed across the study area.}

\item{bin_width}{The width of the climate bins.}
}
\value{
A \code{ccs} object that will be used by \code{\link{fit_pdfsp}}.
}
\description{
Calibrate the distribution of the modern climate space.
}
\examples{
#Extracting the number of taxa recorded in the database
calib_clim_space(sample(0:300/10, 4000, replace=TRUE), 2)

}
