% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Check if the coordinates are correct.}
\usage{
check_coordinates(xmn, xmx, ymn, ymx)
}
\arguments{
\item{xmn, xmx, ymn, ymx}{The coordinates defining the study area.}
}
\value{
Return a set of valid coordinates.
}
\description{
Check if the coordinates are correct.
}
\examples{
check_coordinates(NA, NA, NA, NA)
check_coordinates(-200, 0, 0, 90)
check_coordinates(20, 0, 90, 0)

}
