% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_data.R
\name{cite_climate_data}
\alias{cite_climate_data}
\title{Returns the references associated with the climate data used to fit the \code{pdfs}.}
\usage{
cite_climate_data(x, verbose = TRUE)
}
\arguments{
\item{x}{A \code{\link{crestObj}} produced by one of the \code{\link{crest}},
\code{\link{crest.get_modern_data}}, \code{\link{crest.calibrate}},
\code{\link{crest.reconstruct}} or \code{\link{loo}} functions.}

\item{verbose}{A boolean to print non-essential comments on the terminal
(default \code{TRUE}).}
}
\value{
A list of references to add if the data generated by crestr are published.
}
\description{
Returns the references associated with the climate data used to fit the \code{pdfs}.
}
