% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_GBIF.R
\name{cite_GBIF}
\alias{cite_GBIF}
\title{Returns the citations associated to the GBIF data used to fit the pdfs.}
\usage{
cite_GBIF(x, dbname = "gbif4crest_02", verbose = TRUE)
}
\arguments{
\item{x}{A crestObj produced by one of the crest functions.}

\item{dbname}{The name of the database. Default is gbif4crest_02.}

\item{verbose}{A boolean to print non-essential comments on the terminal (default TRUE).}
}
\description{
Returns the citations associated to the GBIF data used to fit the pdfs.
}
