% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetClimateXY.R
\name{climate_from_xy}
\alias{climate_from_xy}
\title{Extract The climate values associated to a set of coordinates.}
\usage{
climate_from_xy(
  long,
  lat,
  climate = accClimateVariables()[, 2],
  resol = 0.25,
  dbname = "gbif4crest_02"
)
}
\arguments{
\item{long}{The longitude of the site.}

\item{lat}{The latitude of the site.}

\item{climate}{The climate variables to extract the values from. Returns all
possible values by default.}

\item{resol}{The resolution of the target climatology (default 0.25 degrees).}

\item{dbname}{The name of the data source database.}
}
\value{
A data frame containing the climate values.
}
\description{
Extract The climate values associated to a set of coordinates.
}
\examples{
\dontrun{
  climate_from_xy(50, 10, c('bio1', 'ai'))
  climate_from_xy(50, 10)
}

}
