% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDisconnect.R
\name{close_db_connection}
\alias{close_db_connection}
\title{Disconnect the database connection.}
\usage{
close_db_connection(db)
}
\arguments{
\item{db}{An active database connection}
}
\value{
No return value, function called to close the connection to the database.
}
\description{
Disconnect the database connection.
}
\examples{
db <- connect_online()
close_db_connection(db)

}
