% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colour_theme}
\alias{colour_theme}
\title{Returns a vector of colours}
\usage{
colour_theme(n)
}
\arguments{
\item{n}{An index to select the colour theme}
}
\value{
A vector of colours.
}
\description{
Returns a vector of colours
}
\examples{
colour_theme(1)

}
