% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{connect_local_sqlite3}
\alias{connect_local_sqlite3}
\title{Connect to the gbif4crest calibration database}
\usage{
connect_local_sqlite3(dbname = "gbif4crest_02.sqlite3")
}
\arguments{
\item{dbname}{The complete path to the SQLite3 file. The name should end by '.sqlite3'}
}
\value{
An active connection to a database
}
\description{
Connect to the gbif4crest_02 database using a local SQLite3 copy.
}
\examples{
\dontrun{
  db <- connect_online()
}

}
\seealso{
The SQLite3 database can be downloaded from \url{https://figshare.com/articles/dataset/GBIF_for_CREST_database/6743207}.
}
