% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{connect_online}
\alias{connect_online}
\title{Connect to the gbif4crest calibration database}
\usage{
connect_online(
  dbname = "GBIF_v3",
  port = 5432,
  host = "localhost",
  user = "mchevali1",
  password = ""
)
}
\arguments{
\item{dbname}{The name of the database. Default is \code{'gbif4crest_02'}.}

\item{port}{The port to connect to the server. Default is 5432.}

\item{host}{The host of the database server.
Default is \code{'gbif4crest.cvqgy2mnjwtg.eu-west-3.rds.amazonaws.com'}.}

\item{user}{The user name to use to connect. Default is \code{'guestuser'}.}

\item{password}{The password associated with the user name. Default is \code{'pwd12345'}.}
}
\value{
An active connection to a database
}
\description{
Connect to the gbif4crest_02 database by accessing the server on Amazon.
}
\examples{
\dontrun{
db <- connect_online()
}

}
