% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPSE.R
\name{createPSE}
\alias{createPSE}
\title{Returns the citations associated to the GBIF data used to fit the pdfs.}
\usage{
createPSE(taxa, loc = "proxy_species_equivalency.xlsx")
}
\arguments{
\item{taxa}{A list of taxa to include in the PSE file.}

\item{loc}{An absolute or relative path that indicates where the spreadsheet
should be saved. Also used to specify the name of the file. Default:
the file is saved in the working directory under the name
\code{proxy_species_equivalency.xlsx}.}
}
\value{
No return value, called to create a PSE file.
}
\description{
Returns the citations associated to the GBIF data used to fit the pdfs.
}
\examples{
data(crest_ex)
createPSE(taxa = colnames(crest_ex)[-1],
          loc=file.path(tempdir(), 'pse.xlsx')
)

}
