% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.calibrate.R
\name{crest.calibrate}
\alias{crest.calibrate}
\title{Fit the species and proxy pdfs}
\usage{
crest.calibrate(
  x,
  bin_width = rep(1, length(x$parameters$climate)),
  shape = rep("normal", length(x$parameters$climate)),
  npoints = 500,
  geoWeighting = TRUE,
  climateSpaceWeighting = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{a crestObj produced by the crest.climate_space function.}

\item{bin_width}{.}

\item{shape}{.}

\item{npoints}{The number of points to be used to fit the pdfs.}

\item{geoWeighting}{The number of points to be used to fit the pdfs.}

\item{climateSpaceWeighting}{The number of points to be used to fit the pdfs.}

\item{verbose}{A boolean to print non-essential comments on the terminal (default TRUE).}
}
\value{
A crest() object containing the spatial distributions and the climate space
}
\description{
This function fits the climate response of the selected taxa to the selected
climate variables.
}
\examples{
\dontrun{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
x <- crest.get_modern_data(
  pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"),
  selectedTaxa = crest_ex_selection, dbname = "crest_example"
)
x <- crest.calibrate(x,
  geoWeighting = TRUE, climateSpaceWeighting = TRUE,
  bin_width = c(2, 20), shape = c("normal", "lognormal")
)
}

}
