% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.climate_space.R
\name{crest.climate_space}
\alias{crest.climate_space}
\title{Extract the climate space}
\usage{
crest.climate_space(x, dbname = "gbif4crest_02")
}
\arguments{
\item{x}{a crestObj produced by the crest.get_distributions function.}

\item{dbname}{The name of the database. Default is gbif4crest_02.}
}
\value{
A crest() object containing the spatial distributions and the climate space
}
\description{
This function will extract the climate space associated with the species
distributions
}
\examples{
\dontrun{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
x <- crest.get_distributions(taxa.name=colnames(crest_ex)[-1],
                 pse=crest_ex_pse, taxaType = 0,
                 climate = c('bio1', 'bio12'),
                 selectedTaxa = crest_ex_selection, dbname = 'crest_example')
x <- crest.climate_space(x, dbname = 'crest_example')
x
head(x$modelling$climate_space)
dim(x$modelling$climate_space)
}
}
