% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.get_distributions.R
\name{crest.get_distributions}
\alias{crest.get_distributions}
\title{Extract distributions from the database}
\usage{
crest.get_distributions(
  taxa.name,
  pse,
  taxaType,
  climate,
  xmn = -180,
  xmx = 180,
  ymn = -90,
  ymx = 90,
  continents = NA,
  countries = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  minGridCells = 20,
  selectedTaxa = NA,
  dbname = "gbif4crest_02"
)
}
\arguments{
\item{taxa.name}{.}

\item{pse}{.}

\item{taxaType}{.}

\item{climate}{A vectof of the climate variables to extract.}

\item{xmn}{The coordinates defining the study area.}

\item{xmx}{The coordinates defining the study area.}

\item{ymn}{The coordinates defining the study area.}

\item{ymx}{The coordinates defining the study area.}

\item{continents}{A vector of the continent names defining the study area.}

\item{countries}{A vector of the country names defining the study area.}

\item{realms}{A vector of the studied botanical realms defining the study area.}

\item{biomes}{A vector of the studied botanical biomes defining the study area.}

\item{ecoregions}{A vector of the studied botanical ecoregions defining the study area.}

\item{minGridCells}{.}

\item{selectedTaxa}{.}

\item{dbname}{The name of the database. Default is gbif4crest_02.}
}
\value{
A crest() object containing the spatial distributions
}
\description{
This function will extract the distributions of all the species composing each
taxon and return them as a list.
}
\examples{
\dontrun{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
x <- crest.get_distributions(taxa.name=colnames(crest_ex)[-1],
                 pse=crest_ex_pse, taxaType = 0,
                 climate = c('bio1', 'bio12'),
                 selectedTaxa = crest_ex_selection, dbname = 'crest_example')
x
lapply(x$modelling$distributions, head)
}
}
