% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.R
\name{crest}
\alias{crest}
\title{Connect to the gbif4crest database}
\usage{
crest(
  df,
  pse,
  taxaType,
  climate,
  xmn = -180,
  xmx = 180,
  ymn = -90,
  ymx = 90,
  continents = NA,
  countries = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  minGridCells = 20,
  selectedTaxa = NA,
  bin_width = rep(1, length(x$parameters$climate)),
  shape = rep("normal", length(x$parameters$climate)),
  npoints = 500,
  geoWeighting = TRUE,
  climateSpaceWeighting = TRUE,
  presenceThreshold = 0,
  taxWeight = "normalisation",
  uncertainties = c(0.5, 0.95),
  leave_one_out = FALSE,
  verbose = TRUE,
  dbname = "gbif4crest_02"
)
}
\arguments{
\item{df}{.}

\item{pse}{.}

\item{taxaType}{.}

\item{climate}{A vectof of the climate variables to extract.}

\item{xmn}{The coordinates defining the study area.}

\item{xmx}{The coordinates defining the study area.}

\item{ymn}{The coordinates defining the study area.}

\item{ymx}{The coordinates defining the study area.}

\item{continents}{A vector of the continent names defining the study area.}

\item{countries}{A vector of the country names defining the study area.}

\item{realms}{A vector of the studied botanical realms defining the study area.}

\item{biomes}{A vector of the studied botanical biomes defining the study area.}

\item{ecoregions}{A vector of the studied botanical ecoregions defining the
study area.}

\item{minGridCells}{.}

\item{selectedTaxa}{.}

\item{bin_width}{.}

\item{shape}{.}

\item{npoints}{The number of points to be used to fit the pdfs.}

\item{geoWeighting}{The number of points to be used to fit the pdfs.}

\item{climateSpaceWeighting}{The number of points to be used to fit the pdfs.}

\item{presenceThreshold}{.}

\item{taxWeight}{'originalData', 'presence/absence', 'percentages' or
'normalisation'}

\item{uncertainties}{A (vector of) threshold value(s) indicating the error
bars that should be calculated (default both 50 and 95\% ranges).}

\item{leave_one_out}{A boolean to indicate whether the leave one out (loo)
reconstructions should be computed (default FALSE).}

\item{verbose}{A boolean to print non-essential comments on the terminal (default TRUE).}

\item{dbname}{The name of the database. Default is gbif4crest_02.}
}
\value{
The parameters to be used by crest()
}
\description{
Connect to the gbif4crest_02 database by accessing the server on Amazon.
}
\examples{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
recons <- crest(
  df = crest_ex, pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"), bin_width = c(2, 20),
  shape = c("normal", "lognormal"),
  selectedTaxa = crest_ex_selection, dbname = "crest_example",
  leave_one_out = TRUE
)
plot(recons)
plot_loo(recons)

}
