% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.reconstruct.R
\name{crest.reconstruct}
\alias{crest.reconstruct}
\title{Fit the species and proxy pdfs}
\usage{
crest.reconstruct(
  x,
  df,
  presenceThreshold = 0,
  taxWeight = "normalisation",
  skip_for_loo = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A crestObj produced by the crest.fit_pdfs function.}

\item{df}{.}

\item{presenceThreshold}{.}

\item{taxWeight}{'originalData', 'presence/absence', 'percentages' or 'normalisation'}

\item{skip_for_loo}{A boolean that tells the loo() functiont to skip parts
and fasten the process. Not for users.}

\item{verbose}{A boolean to print non-essential comments on the terminal (default TRUE).}
}
\value{
A crest() object containing the reconstructions and all the associated data.
}
\description{
This function fits the climate response of the selected taxa to the selected
climate variables.
}
\examples{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
x <- crest.get_modern_data(
  pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"),
  taxa.name = colnames(crest_ex)[-1],
  selectedTaxa = crest_ex_selection, dbname = "crest_example"
)
x <- crest.calibrate(x,
  geoWeighting = TRUE, climateSpaceWeighting = TRUE,
  bin_width = c(2, 20), shape = c("normal", "lognormal")
)
x <- crest.reconstruct(x, crest_ex)
plot(x)
}
