% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crest.simplify}
\alias{crest.simplify}
\title{Simplify a crestObj into a dataframe.}
\usage{
crest.simplify(x, optima = TRUE)
}
\arguments{
\item{x}{A \code{\link{crestObj}} produced by the \code{\link{crest}},
\code{\link{crest.reconstruct}} or \code{\link{loo}} functions.}

\item{optima}{A boolean to indicate whether to plot the optimum (\code{TRUE})
or the mean (\code{FALSE}) estimates.}
}
\value{
A dataframe with the age/depth of each sample and all the best
reconstructed values.
}
\description{
Simplify a crestObj with reconstructed values into a dataframe.
}
\examples{
head(crest.simplify(reconstr))

}
