% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crest_ex}
\alias{crest_ex}
\title{Example dataset to run the CREST method for the first time.}
\format{
A data frame with 20 rows (samples) and 8 columns (1 column for the
age and one for each of the 7 taxa):
\describe{
\item{Age:}{Age of each sample}
\item{Taxon1:}{Percentage of Taxon1 in each sample.}
\item{Taxon2:}{Percentage of Taxon2 in each sample.}
\item{Taxon3:}{Percentage of Taxon3 in each sample.}
\item{Taxon4:}{Percentage of Taxon4 in each sample.}
\item{Taxon5:}{Percentage of Taxon5 in each sample.}
\item{Taxon6:}{Percentage of Taxon6 in each sample.}
\item{Taxon7:}{Percentage of Taxon7 in each sample.}
}
}
\usage{
crest_ex
}
\description{
A dataset containing 20 randomly generated pollen samples for 7 pollen taxa.
}
\keyword{datasets}
