% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crest_ex_pse}
\alias{crest_ex_pse}
\title{Example dataset to Extract data from the example database.}
\format{
A data frame with 7 rows (taxa) and 5 columns (taxonomy description):
\describe{
\item{Level:}{An integer indicating the taxonomic resolution
(1 family, 2 genus, 3 species, 4 or higher ignore taxon)}
\item{Family:}{The family corresponding to the ProxyName}
\item{Genus:}{The genus corresponding to the ProxyName}
\item{Species:}{The species corresponding to the ProxyName}
\item{ProxyName:}{The names of the observed proxies, as reported in the
main data file}
}
}
\usage{
crest_ex_pse
}
\description{
A database indicating the taxonomy of the example proxies.
}
\keyword{datasets}
