% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crest_ex_selection}
\alias{crest_ex_selection}
\title{Example dataset to associate taxa with climate variables.}
\format{
A data frame with 7 rows (taxa) and 2 columns (climate variables):
\describe{
\item{bio1:}{The first variable to reconstruct (mean annual temperature)}
\item{bio12:}{The second variable to reconstruct (annual precipitation)}
}
}
\usage{
crest_ex_selection
}
\description{
A data frame indicating the taxa that should be used to reconstruct each
climate variable (1s in the matrix) and those who should be excluded (0s).
}
\keyword{datasets}
