% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{dbDisconnect}
\alias{dbDisconnect}
\title{Disconnect the database connection.}
\usage{
dbDisconnect(db)
}
\arguments{
\item{db}{An active database connection}
}
\description{
Disconnect the database connection.
}
\examples{
\dontrun{
db <- connect_online()
dbDisconnect(db)
}
}
