% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRequest.R
\name{dbRequest}
\alias{dbRequest}
\title{Connect to the gbif4crest database}
\usage{
dbRequest(request)
}
\arguments{
\item{request}{A SQL request to be executed.}
}
\value{
The result of the request.
}
\description{
Connect to the gbif4crest_02 database by accessing the server on Amazon.
}
\examples{
dbRequest("SELECT count(*) FROM taxa")
southAfricaTaxa <- dbRequest(paste0("SELECT taxa.* ",
    "FROM taxa, distrib_qdgc, geo_qdgc ",
    "WHERE taxa.taxonid=distrib_qdgc.taxonid ",
    "AND   distrib.qdgc.latitude=geo_qdgc.latitude ",
    "AND   distrib_qdgc.longitude=geo_qdgc.longitude ",
    "AND geo_qdgc.countryname='South Africa'"))
}
