% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreateSubset.R
\name{dbSubset}
\alias{dbSubset}
\title{Create a subset of the global calibration dataset}
\usage{
dbSubset(
  taxaType,
  xmn = NA,
  xmx = NA,
  ymn = NA,
  ymx = NA,
  dbname = "gbif4crest_02",
  out = "gbif4crest_reduced",
  verbose = TRUE
)
}
\arguments{
\item{taxaType}{A numerical index (between 1 and 6) to define the type of
palaeoproxy used: 1 for plants, 2 for beetles, 3 for chironomids,
4 for foraminifers, 5 for diatoms and 6 for rodents. The example
dataset uses taxaType=0 (pseudo-data). Default is 1.}

\item{xmn, xmx, ymn, ymx}{The coordinates defining the study area.}

\item{dbname}{The name of the database. Default is \code{'gbif4crest_02'} and
data will be extracted from the online database. The SQLite3 version
of the database can also be used here by providing the complete path
to a file ending by \code{.sqlite3}, e.g. \code{/path/to/file/gbif4crest_02.sqlite3}}

\item{out}{The name or path of the new dataset}

\item{verbose}{A boolean to print non-essential comments on the terminal
(default \code{TRUE}).}
}
\description{
Create a subset of the global calibration dataset
}
\examples{
\dontrun{
  dbSubset(2, xmn=0, xmx=15, ymn=0, ymx=15, out='example.sqlite3')
}

}
\seealso{
The full SQLite3 database can be downloaded from \url{https://figshare.com/articles/dataset/GBIF_for_CREST_database/6743207}.
}
