% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{.ecoid2names}
\alias{.ecoid2names}
\title{Return biogeographical units associated with a ecoID.}
\usage{
.ecoid2names(ids, realm = 1, dbname = "gbif4crest_02")
}
\arguments{
\item{ids}{A vector of IDs (mari_ecoid or terr_ecoid)}

\item{realm}{An index that says if the ids are from terrestrial (realm=1;
default) or from marine (realm=2) settings.}

\item{dbname}{The database to use.}
}
\value{
A list of names.
}
\description{
Return biogeographical units associated with a ecoID.
}
