% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{.geopoid2names}
\alias{.geopoid2names}
\title{Return geopolitical units associated with a geopoID.}
\usage{
.geopoid2names(ids, realm = 1, dbname = "gbif4crest_02")
}
\arguments{
\item{ids}{A vector of IDs (countryid or oceanid)}

\item{realm}{An index that says if the ids are from terrestrial (realm=1;
default) or from marine (realm=2) settings.}

\item{dbname}{The database to use.}
}
\value{
A list of names.
}
\description{
Return geopolitical units associated with a geopoID.
}
\keyword{internal}
