% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{.getRealmNames}
\alias{.getRealmNames}
\title{Return the list of the realms and associated biomes and ecoregions.}
\usage{
.getRealmNames(ecoregion = TRUE, dbname = "gbif4crest_02")
}
\arguments{
\item{ecoregion}{A boolean to choose whether to get the ecoregions names.}

\item{dbname}{The name of the database. Default is gbif4crest_02.}
}
\value{
A list with elements that correspond to the biomes (and possibly
ecoregions) of each realm.
}
\description{
Return the list of the realms and associated biomes and ecoregions.
}
