% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeTaxa.R
\name{excludeTaxa}
\alias{excludeTaxa}
\title{Excludes the list of taxa to the reconstructions.}
\usage{
excludeTaxa(x, taxa, climate)
}
\arguments{
\item{x}{A crestObj produced by one of the crest functions.}

\item{taxa}{A vector of taxa to Exclude.}

\item{climate}{A vector of climate variables to unlink the taxa with.}
}
\description{
Excludes the list of taxa to the reconstructions.
}
