% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export the results}
\usage{
export(
  x,
  sitename = x$misc$site_info$sitename,
  climate = x$parameters$climate,
  loc = getwd(),
  as.csv = FALSE,
  fullPosterior = FALSE,
  loo = FALSE,
  weights = FALSE,
  pdfs = FALSE
)
}
\arguments{
\item{x}{The name, age or depth of the rows of df (the samples).}

\item{sitename}{The name of the site (default: x$misc$site_info$sitename)}

\item{climate}{The climate data to export.}

\item{loc}{The path where to export the data (default: working directory)}

\item{as.csv}{Boolean to indicate if the data should be exported as csv (TRUE) or xlsx (FALSE, default)}

\item{fullPosterior}{A boolean to export the climate posterior probability (default FALSE)}

\item{loo}{A boolean to export the leave-one-out data if they exist (default FALSE)}

\item{weights}{A boolean to export the weights derived from the percentages (default FALSE)}

\item{pdfs}{A boolean to export the taxa's pdfs (default FALSE)}
}
\description{
Export the results generated by the reconstruction
}
\examples{
\dontrun{
  data(crest_ex)
  data(crest_ex_pse)
  data(crest_ex_selection)
  recons <- crest(
      df = crest_ex, pse = crest_ex_pse, taxaType = 0,
    site_info = c(7.5, 7.5),
    climate = c("bio1", "bio12"), bin_width = c(2, 20),
    shape = c("normal", "lognormal"),
    selectedTaxa = crest_ex_selection, dbname = "crest_example",
    leave_one_out = TRUE
  )
  export(recons, sitename='test3', fullPosterior=TRUE, weights=TRUE, loo=TRUE)
}

}
