% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_locid}
\alias{f_locid}
\title{Calculates a unique identifier from coordinates.}
\usage{
f_locid(long, lat, resol)
}
\arguments{
\item{long}{The longitude of the grid cell of interest}

\item{lat}{The latitude of the grid cell of interest}

\item{resol}{The spatial resolution of the database to target (1/12 for gbif4crest_03)}
}
\value{
A unique integer describing the cell.
}
\description{
Calculates a unique identifier from coordinates.
}
\examples{
f_locid(0, 0, 1/12)

}
