% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find.original.name}
\alias{find.original.name}
\title{Returns the name of the function argument in the global environment}
\usage{
find.original.name(x)
}
\arguments{
\item{x}{The function argument}
}
\value{
The name of the function argument in the global environment.
}
\description{
Returns the name of the function argument in the global environment
}
