% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit_pdfsp.R
\name{fit_xrange}
\alias{fit_xrange}
\title{Define teh climate gradient to fit the pdfs.}
\usage{
fit_xrange(ccs, shape, bin_width, npoints = 500)
}
\arguments{
\item{ccs}{A \code{ccs} object returned by \code{\link{calib_clim_space}}.}

\item{shape}{The shape of the species pdfs. Use 'normal' or 'lognormal'.}

\item{bin_width}{The width of the climate bins.}

\item{npoints}{The number of points to be used to fit the pdfs.}
}
\value{
A regularly spaced climate gradient with \code{npoints} points.
}
\description{
Define teh climate gradient to fit the pdfs.
}
\examples{
# Creating one randomised climate space
climate_space <- sample(0:300 / 10, 4000, replace = TRUE)
ccs <- calib_clim_space(climate_space, 2)
xrange <- fit_xrange(ccs, "normal", 2)
head(xrange)
}
