% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_taxa_type}
\alias{get_taxa_type}
\title{Returns the taxa type corresponding to the index.}
\usage{
get_taxa_type(taxaType)
}
\arguments{
\item{taxaType}{An integer between 0 and 6}
}
\value{
Returns the taxa type corresponding to the index.
}
\description{
Returns the taxa type corresponding to the index.
}
