% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{getBasinNames}
\alias{getBasinNames}
\title{Return the list of oceans and seas.}
\usage{
getBasinNames(dbname = "gbif4crest_02")
}
\arguments{
\item{dbname}{The name of the database. Default is \code{'gbif4crest_02'}.}
}
\value{
A list where each element is a vector of corresponding ocean/sea names.
}
\description{
Return the list of oceans and seas.
}
\keyword{internal}
