% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{getCountryNames}
\alias{getCountryNames}
\title{Return the list of the continents and associated countries.}
\usage{
getCountryNames(dbname = "gbif4crest_02")
}
\arguments{
\item{dbname}{The name of the database. Default is \code{'gbif4crest_02'}.}
}
\value{
A list where each element is a vector of corresponding country names.
}
\description{
Return the list of the continents and associated countries.
}
\keyword{internal}
