% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getResol}
\alias{getResol}
\title{Calculates a unique identifier from coordinates.}
\usage{
getResol(crest)
}
\arguments{
\item{crest}{A crestObj object.}
}
\value{
A unique integer describing the cell.
}
\description{
Calculates a unique identifier from coordinates.
}
\examples{
getResol(reconstr)

}
