% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTaxaTypeFromTaxID}
\alias{getTaxaTypeFromTaxID}
\title{Returns the taxa type corresponding to the taxID.}
\usage{
getTaxaTypeFromTaxID(taxID)
}
\arguments{
\item{taxID}{An integer between 0 and 6}
}
\value{
Returns the taxa type ID corresponding to the taxon ID.
}
\description{
Returns the taxa type corresponding to the taxID.
}
