% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identifyDatabase}
\alias{identifyDatabase}
\title{Identify the calibration database used.}
\usage{
identifyDatabase(dbname)
}
\arguments{
\item{dbname}{A functional crestObj or a database name.}
}
\value{
A string uniquely characterising the database used. Possible values
are 'privateDB', 'exampleDB', 'gbif4crest_02', and 'gbif4crest_03'.
}
\description{
Identify the calibration database used.
}
\examples{
identifyDatabase(reconstr)

}
