% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.crestObj}
\alias{is.crestObj}
\title{Test if x is a crestObj.}
\usage{
is.crestObj(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
\code{TRUE} (x is a crestObj) or \code{FALSE} (not a crestObj).
}
\description{
Test if x is a crestObj.
}
