% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_database_gbif}
\alias{is_database_gbif}
\title{Check if the database is one of mine.}
\usage{
is_database_gbif(dbname)
}
\arguments{
\item{dbname}{A functional crestObj or a database name.}
}
\value{
A boolean indicating if the database if of the gbif4crest Family
or the example database (TRUE). FALSE otherwise.
}
\description{
Check if the database is one of mine.
}
\examples{
is_database_gbif(reconstr)

}
