% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isColourStr.R
\name{isColourStr}
\alias{isColourStr}
\title{Test if R can interpret a string as a colour}
\usage{
isColourStr(col)
}
\arguments{
\item{col}{The string to be tested.}
}
\value{
A boolean value, TRUE if col is a valid colour, FALSE otherwise
}
\description{
Test if R can interpret a string as a colour
}
\examples{
isColourStr('black')
isColourStr('blakc')

}
