% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_one_out.R
\name{loo}
\alias{loo}
\title{Connect to the gbif4crest database}
\usage{
loo(x)
}
\arguments{
\item{x}{a crestObj produced by the crest.reconstruct() or crest() functions.}
}
\value{
A crest() object containing the reconstructions and all the associated data.
}
\description{
Connect to the gbif4crest_02 database by accessing the server on Amazon.
}
\examples{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
recons <- crest(
  df = crest_ex, pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"), bin_width = c(2, 20),
  shape = c("normal", "lognormal"),
  selectedTaxa = crest_ex_selection, dbname = "crest_example"
)
recons <- loo(recons)
recons$reconstructions$bio12$loo
plot_loo(recons)

}
