% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M1}
\alias{M1}
\title{A shapefile of the world's country borders. Use 'M1 <- terra::unwrap(M1)' to
render the object usable in R.}
\format{
An object of class \code{PackedSpatVector} of length 1.
}
\usage{
M1
}
\description{
A shapefile of the world's country borders. Use 'M1 <- terra::unwrap(M1)' to
render the object usable in R.
}
\keyword{datasets}
