% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Wrapper function of to add transparency to a colour.}
\usage{
makeTransparent(colour, alpha)
}
\arguments{
\item{colour}{A R colour}

\item{alpha}{A value between 0 and 1 that defines the transparency
0 for full transparency and 1 for no transparency}
}
\description{
Add transparency to the selected colours.
}
\examples{
makeTransparent('black',0.5)
makeTransparent('black',1:10/10)
makeTransparent(rainbow(10), 1:10/10)
}
