% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2percentages.R
\name{meanPositiveValues}
\alias{meanPositiveValues}
\title{Convert data into presence/absence data.}
\usage{
meanPositiveValues(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The average of all the positive values.
}
\description{
Convert data into presence/absence data.
}
\examples{
meanPositiveValues(-10:10)
}
