% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.climateSpace.R
\name{plot_climateSpace}
\alias{plot_climateSpace}
\title{Plot the studied climate space.}
\usage{
plot_climateSpace(
  x,
  save = FALSE,
  loc = "Climate_space.pdf",
  width = 7.48,
  height = min(9, 3 * length(x$parameters$climate)),
  y0 = 0.3
)
}
\arguments{
\item{x}{A crestObj generated by either the crest.calibrate(),
crest.reconstrut() or crest() functions.}

\item{save}{A boolean to indicate if the diagram shoud be saved as a pdf file.
Default is FALSE.}

\item{loc}{An absolute or relative path that indicates where the diagram
should be saved. Also used to specify the name of the file. Default:
the file is saved in the working directory under the name
Climate_space.pdf.}

\item{width}{The width of the output file in inches (default 7.48in ~ 19cm).}

\item{height}{The height of the output file in inches (default 3in ~ 7.6cm
per variables).}

\item{y0}{The space to allocate to each title (default 0.3in ~ 0.76 cm)}
}
\description{
Plot the studied climate space.
}
\examples{
\dontrun{
  data(crest_ex_pse)
  data(crest_ex_selection)
  x <- crest.get_modern_data(
    pse = crest_ex_pse, taxaType = 0,
    climate = c("bio1", "bio12"),
    selectedTaxa = crest_ex_selection, dbname = "crest_example"
  )
  x <- crest.calibrate(x,
    geoWeighting = TRUE, climateSpaceWeighting = TRUE,
    bin_width = c(2, 20), shape = c("normal", "lognormal")
  )
  plot_climateSpace(x)
}

}
