% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.combinedPDFS.R
\name{plot_combinedPDFs}
\alias{plot_combinedPDFs}
\title{Plot representing how the pdfs combine to produce the reconstruction.}
\usage{
plot_combinedPDFs(
  x,
  samples = 1:length(x$inputs$x),
  climate = x$parameters$climate[1],
  optima = TRUE,
  save = FALSE,
  loc = "samplePDFs.pdf",
  width = 7.48,
  height = 5
)
}
\arguments{
\item{x}{A crestObj generated by the crest.reconstruct() or crest() functions.}

\item{samples}{The list of samples for which the plot should be plotted. All
samples will be plotted by default.}

\item{climate}{The climate variable to use to plot the variable. Default is
first variable (x$parameters$climate[1]).}

\item{optima}{A boolean to indicate whether to plot the optimum (TRUE) or the
mean (FALSE) estimates.}

\item{save}{A boolean to indicate if the diagram shoud be saved as a pdf file.
Default is FALSE.}

\item{loc}{An absolute or relative path that indicates where the diagram
should be saved. Also used to specify the name of the file. Default:
the file is saved in the working directory under the name
samplePDFs.pdf.}

\item{width}{The width of the output file in inches (default 7.48in ~ 19cm).}

\item{height}{The height of the output file in inches (default 3in ~ 7.6cm
per variables).}
}
\description{
Plot representing how the pdfs combine to produce the reconstruction.
}
\examples{
\dontrun{
  data(crest_ex)
  data(crest_ex_pse)
  data(crest_ex_selection)
  recons <- crest(
    df = crest_ex, pse = crest_ex_pse, taxaType = 0,
    climate = c("bio1", "bio12"), bin_width = c(2, 20),
    shape = c("normal", "lognormal"),
    selectedTaxa = crest_ex_selection, dbname = "crest_example",
    leave_one_out = FALSE
   )
  plot_combinedPDFs(recons, samples=1:12, climate='ai')
}

}
