% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loo.R
\name{plot_loo}
\alias{plot_loo}
\title{Plot stratigraphic data as polygons or barplots.}
\usage{
plot_loo(
  x,
  save = FALSE,
  filename = "Diagram_loo",
  width = 3.54,
  height = 9,
  yax_incr = NA,
  bar_width = 1,
  xlim = NA,
  tickAtSample = FALSE,
  col_pos = "black",
  col_neg = "white",
  title = NA
)
}
\arguments{
\item{x}{A data frame of the data to plot (first column with age or depth)
and the taxa in the following columns. x can also be a crestObj.}

\item{save}{A boolean to indicate if the diagram shoud be saved as a pdf file.
Default is FALSE.}

\item{filename}{An absolute or relative path that indicates where the diagram
should be saved. Also used to specify the name of the file. Default:
the file is saved in the working directory under the name Diagram.pdf.}

\item{width}{The width of the output file in inches (default 3.54in ~ 9cm).}

\item{height}{The height of the output file in inches (default 9in ~ 23cm).}

\item{yax_incr}{Graphical parameters describing the increment size on the y-axis (default 5).}

\item{bar_width}{Width of the bars of the barplot (default 1).}

\item{xlim}{The range covered by the x-axis. Canbe adjusted to get round numbers on the x-ais. If smaller than the range overed by the data, the data will be truncated (default: range of the data).}

\item{tickAtSample}{Boolean that indicates whether a tick mark should be added on the x-axis at the location of each sample (default TRUE).}

\item{col_pos}{Graphical parameter for the barplot. Colour of all the positive values (default black).}

\item{col_neg}{Graphical parameter for the barplot. Colour of all the negative values (default white).}

\item{title}{Name to be added on top of the plot (default NA).}
}
\description{
This function plots stratigraphic data either as polygons or bars.
}
\examples{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
recons <- crest(
  df = crest_ex, pse = crest_ex_pse, taxaType = 0,
  climate = c("bio1", "bio12"), bin_width = c(2, 20),
  shape = c("normal", "lognormal"),
  selectedTaxa = crest_ex_selection, dbname = "crest_example"
)
recons <- loo(recons)
\dontrun{
plot_loo(recons, yax_incr=c(0.5, 50), bar_width=0.8,
         col_pos=c('blue','cornflowerblue'), col_neg=c('red', 'goldenrod3'))
}

}
