% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.speciesCharacteristics.R
\name{plot_taxaCharacteristics}
\alias{plot_taxaCharacteristics}
\title{Plot the studied climate space.}
\usage{
plot_taxaCharacteristics(
  x,
  taxanames = x$inputs$taxa.name,
  save = FALSE,
  loc = "taxaCharacteristics.pdf",
  width = 7.48,
  w0 = 0.3,
  height = min(9, 3 * length(x$parameters$climate)),
  h0 = 0.3
)
}
\arguments{
\item{x}{A crestObj generated by either the crest.calibrate(),
crest.reconstrut() or crest() functions.}

\item{taxanames}{A list of taxa to use for the plot (default is all the
recorded taxa).}

\item{save}{A boolean to indicate if the diagram shoud be saved as a pdf file.
Default is FALSE.}

\item{loc}{An absolute or relative path that indicates where the diagram
should be saved. Also used to specify the name of the file. Default:
the file is saved in the working directory under the name
Climate_space.pdf.}

\item{width}{The width of the output file in inches (default 7.48in ~ 19cm).}

\item{w0}{The width of the left column with the names.}

\item{height}{The height of the output file in inches (default 3in ~ 7.6cm
per variables).}

\item{h0}{The vertical space used for the x-axes.}
}
\description{
Plot the studied climate space.
}
\examples{
\dontrun{
  data(crest_ex_pse)
  data(crest_ex_selection)
  x <- crest.get_modern_data(
    pse = crest_ex_pse, taxaType = 0,
    climate = c("bio1", "bio12"),
    selectedTaxa = crest_ex_selection, dbname = "crest_example"
  )
  x <- crest.calibrate(x,
    geoWeighting = TRUE, climateSpaceWeighting = TRUE,
    bin_width = c(2, 20), shape = c("normal", "lognormal")
  )
  plot_taxaCharacteristics(x)
}

}
