% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.PSE_log.R
\name{PSE_log}
\alias{PSE_log}
\title{Displays the log of the proxy-species association}
\usage{
PSE_log(x)
}
\arguments{
\item{x}{A \code{\link{crestObj}} generated by one of the following functions:
\code{\link{crest.get_modern_data}}, \code{\link{crest.set_modern_data}},
\code{\link{crest.calibrate}}, \code{\link{crest.reconstruct}},
\code{\link{loo}}, or \code{\link{crest}} functions.}
}
\description{
Displays the log of the proxy-species association performed by the
crest.get_modern_data() or crest.set_modern_data()
}
\examples{

PSE_log(reconstr)

}
