% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonComposition.R
\name{taxonComposition}
\alias{taxonComposition}
\title{Return the size of the distribution of each composing species of each taxon}
\usage{
taxonComposition(x, taxanames = x$input$taxa.name)
}
\arguments{
\item{x}{A \code{\link{crestObj}} generated by either the \code{\link{crest.calibrate}},
\code{\link{crest.reconstruct}} or \code{\link{crest}} functions.}

\item{taxanames}{A list of taxa to use (default is all the
recorded taxa).}
}
\value{
A list with the number of unique occurrences for each composing species
}
\description{
Return the size of the distribution of each composing species of each taxon
}
\examples{

\dontrun{
  data(reconstr)
  taxonComposition(reconstr)
 }

}
