% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTestConn.R
\name{testConnection}
\alias{testConnection}
\title{Test if a connection can be established with the calibration data}
\usage{
testConnection(dbname = "gbif4crest_02")
}
\arguments{
\item{dbname}{The name of the data source database.}
}
\value{
\code{TRUE} if the connection can be established, else \code{FALSE}.
}
\description{
Test if a connection can be established with the calibration data
}
