% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{accBasinNames}
\alias{accBasinNames}
\title{Return the list of oceans and seas.}
\usage{
accBasinNames(basin = NA)
}
\arguments{
\item{basin}{A name of basin. Default is \code{NA} and returns a list
of all the accepted names.}
}
\value{
A list of accepted names.
}
\description{
Return the list of oceans and seas.
}
\examples{
accBasinNames()
accBasinNames('Indian Ocean')
}
\seealso{
\url{https://www.marineregions.org/downloads.php}
}
