% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{accCountryNames}
\alias{accCountryNames}
\title{Return the list of the continents and associated countries.}
\usage{
accCountryNames(continent = NA)
}
\arguments{
\item{continent}{A name of continent. Default is \code{NA} and returns a list
of all the country names sorted by continent.}
}
\value{
A list where each element is a vector of corresponding country names.
}
\description{
Return the list of the continents and associated countries.
}
\examples{
accCountryNames()
accCountryNames('Europe')
}
\seealso{
\url{https://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries/}
}
