% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFixedNames.R
\name{accRealmNames}
\alias{accRealmNames}
\title{Return the list of the realms and associated biomes and ecoregions.}
\usage{
accRealmNames(ecoregion = TRUE)
}
\arguments{
\item{ecoregion}{A boolean to choose whether to get the ecoregions names.}
}
\value{
A list with elements that correspond to the biomes (and possibly
ecoregions) of each realm.
}
\description{
Return the list of the realms and associated biomes and ecoregions.
}
\examples{
accRealmNames()
}
