% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestInit.R
\name{crest.init}
\alias{crest.init}
\title{Connect to the gbif4crest database}
\usage{
crest.init(df, pse, taxaType)
}
\arguments{
\item{df}{.}

\item{pse}{.}

\item{taxaType}{.}
}
\value{
The parameters to be used by crest()
}
\description{
Connect to the gbif4crest_02 database by accessing the server on Amazon.
}
\examples{
\dontrun{
db <- connect_online()
}
}
