% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.reconstruct.R
\name{crest.reconstruct}
\alias{crest.reconstruct}
\title{Fit the species and proxy pdfs}
\usage{
crest.reconstruct(
  x,
  df,
  selectedTaxa = x$inputs$selectedTaxa,
  presenceThreshold = 0,
  taxWeight = "normalisation"
)
}
\arguments{
\item{x}{a crestObj produced by the crest.fit_pdfs function.}

\item{df}{.}

\item{selectedTaxa}{.}

\item{presenceThreshold}{.}

\item{taxWeight}{'originalData', 'presence/absence', 'percentages' or 'normalisation'}
}
\value{
A crest() object containing the reconstructions and all the associated data.
}
\description{
This function fits the climate response of the selected taxa to the selected
climate variables.
}
\examples{
data(crest_ex)
data(crest_ex_pse)
data(crest_ex_selection)
x <- crest.get_distributions(taxa.name=colnames(crest_ex)[-1],
                 pse=crest_ex_pse, taxaType = 0,
                 climate = c('bio1', 'bio12'),
                 selectedTaxa = crest_ex_selection, dbname = 'crest_example')
x <- crest.climate_space(x, dbname = 'crest_example')
x <- crest.fit_pdfs(x, geoWeighting=TRUE, climateSpaceWeighting=TRUE,
                 bin_width = c(2, 20), shape = c('normal', 'lognormal'))
x <- crest.reconstruct(x, crest_ex, selectedTaxa = crest_ex_selection)
plot(x)

}
