% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestS3.R
\name{crestObj}
\alias{crestObj}
\title{Create a crest() object.}
\usage{
crestObj(
  taxa.name,
  pse,
  taxaType,
  climate,
  xmn,
  xmx,
  ymn,
  ymx,
  continents,
  countries,
  realms,
  biomes,
  ecoregions,
  df = NA,
  x = NA,
  x.name = "",
  minGridCells = 20,
  bin_width = rep(1, length(climate)),
  shape = rep("normal", length(climate)),
  npoints = 500,
  geoWeighting = TRUE,
  climateSpaceWeighting = TRUE,
  selectedTaxa = NA,
  presenceThreshold = 0,
  taxWeight = "normalisation",
  uncertainties = c(0.5, 0.95)
)
}
\arguments{
\item{taxa.name}{.}

\item{pse}{.}

\item{taxaType}{.}

\item{climate}{A vectof of the climate variables to extract.}

\item{xmn, xmx, ymn, ymx}{The coordinates defining the study area.}

\item{continents}{A vector of the continent names defining the study area.}

\item{countries}{A vector of the country names defining the study area.}

\item{realms}{A vector of the studied botanical realms defining the study area.}

\item{biomes}{A vector of the studied botanical biomes defining the study area.}

\item{ecoregions}{A vector of the studied botanical ecoregions defining the
study area.}

\item{df}{.}

\item{x}{.}

\item{x.name}{.}

\item{minGridCells}{.}

\item{bin_width}{.}

\item{shape}{.}

\item{npoints}{The number of points to be used to fit the pdfs.}

\item{geoWeighting}{The number of points to be used to fit the pdfs.}

\item{climateSpaceWeighting}{The number of points to be used to fit the pdfs.}

\item{selectedTaxa}{.}

\item{presenceThreshold}{.}

\item{taxWeight}{'originalData', 'presence/absence', 'percentages' or
'normalisation'}

\item{uncertainties}{A (vector of) threshold value(s) indicating the error
bars that should be calculated (default both 50 and 95\% ranges).}
}
\value{
A CREST object that is used to store data and information for reconstructing climate
}
\description{
Creates a crest() object with all default parameters.
}
