% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fit_pdfsp.R
\name{fit_pdfsp}
\alias{fit_pdfsp}
\title{Fit the species pdfs.}
\usage{
fit_pdfsp(climate, ccs, bin_width, shape, xrange, use_ccs = TRUE)
}
\arguments{
\item{climate}{A vector of climatic values where the species is present.}

\item{ccs}{A \code{ccs} object returned by \code{\link{calib_clim_space}}.}

\item{bin_width}{The width of the climate bins.}

\item{shape}{The shape of the species pdfs. Use 'normal' or 'lognormal'.}

\item{xrange}{The climate gradient upon which the pdf with be defined.}

\item{use_ccs}{Boolean to indicate if the pdfsp should be corrected by the
distributin of the modern climate space}
}
\value{
The pdf of the species.
}
\description{
Fit the species pdfs.
}
\examples{
#Creating one randomised species
climate_species <- round(stats::rnorm(50, 15, 2),1)

#Creating one randomised climate space
climate_space <- base::sample(0:300/10, 4000, replace=TRUE)

ccs <- calib_clim_space(climate_space, 2)
xrange <- fit_xrange(ccs, 'normal', 2)
pdfsp <- fit_pdfsp(climate_species, ccs, 2, 'normal', xrange)
plot(xrange, pdfsp, type='l')

#Testing that the area under the curve is equal to 1.
sum(pdfsp * (xrange[2] - xrange[1])) == 1
}
