% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetDistribTaxa.R
\name{getDistribTaxa}
\alias{getDistribTaxa}
\title{Extract taxonID(s) corresponding to the taxonomic description}
\usage{
getDistribTaxa(
  taxIDs,
  climate = NA,
  xmn = -180,
  xmx = 180,
  ymn = -90,
  ymx = 90,
  continents = NA,
  countries = NA,
  realms = NA,
  biomes = NA,
  ecoregions = NA,
  dbname = "gbif4crest_02"
)
}
\arguments{
\item{taxIDs}{A vector of accepted Taxa IDs (as returned by \code{\link{getTaxonID}}).}

\item{climate}{A vectof of the climate variables to extract.}

\item{xmn, xmx, ymn, ymx}{The coordinates defining the study area.}

\item{continents}{A vector of the continent names defining the study area.}

\item{countries}{A vector of the country names defining the study area.}

\item{realms}{A vector of the studied botanical realms defining the study area.}

\item{biomes}{A vector of the studied botanical biomes defining the study area.}

\item{ecoregions}{A vector of the studied botanical ecoregions defining the study area.}

\item{dbname}{The name of the database. Default is gbif4crest_02.}
}
\value{
A matrix of occurrence records with the associated climate.
}
\description{
Extract all possible TaxonIDs corresponding to the provided taxonomical
description, which can be at the family, the genus or the species levels.
}
\examples{
taxIDs <- getTaxonID("Zamiaceae", "Ceratozamia")
distrib <- getDistribTaxa(taxIDs, "bio1", -90, 90, -90, 90,
  continents = "Europe",
  countries = c("Germany", "Netherlands", "Sweden"),
  realms = "Palaearctic"
)
distrib
}
\seealso{
\code{\link{getTaxonID}} for taxIDs, \code{\link{accClimateVariables}}
for a list of accepted climate variable names, \code{\link{accContinentNames}}
for a list of accepted continent and country names, \code{\link{accRealmNames}}
for a list of accepted realm, biome and ecoregion names.
}
