% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeTaxa.R
\name{includeTaxa}
\alias{includeTaxa}
\title{Includes the list of taxa to the reconstructions.}
\usage{
includeTaxa(x, taxa, climate)
}
\arguments{
\item{x}{A crestObj produced by one of the crest functions.}

\item{taxa}{A vector of taxa to include.}

\item{climate}{A vector of climate variables to link the taxa with.}
}
\description{
Includes the list of taxa to the reconstructions.
}
