% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isColourStr.R
\name{isColourStr}
\alias{isColourStr}
\title{Test if \code{R} can interpret a string as a colour}
\usage{
isColourStr(col)
}
\arguments{
\item{col}{The string to be tested.}
}
\value{
A boolean value, \code{TRUE} if col is a valid colour, \code{FALSE} otherwise
}
\description{
Test if \code{R} can interpret a string as a colour
}
\examples{
isColourStr('black')
isColourStr('blakc')

}
