% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M1}
\alias{M1}
\title{A shapefile of the world's country borders.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 252 rows and 1 columns.
}
\usage{
M1
}
\description{
A shapefile of the world's country borders.
}
\keyword{datasets}
