% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestS3.R
\name{plot.crestObj}
\alias{plot.crestObj}
\title{Plot the reconstructions.}
\usage{
\method{plot}{crestObj}(
  x,
  climate = x$parameters$climate,
  errors = x$parameters$uncertainties,
  optima = TRUE,
  xlim = NA,
  ylim = NA,
  save = FALSE,
  plot = TRUE,
  loc = getwd(),
  ...
)
}
\arguments{
\item{x}{A crestObj produced by the crest.reconstruct() or crest() functions.}

\item{optima}{A boolean to indicate if the optima (TRUE) or the mean (FALSE)
should be plotted.}
}
\description{
Plot the reconstructions and their uncertainties
}
