% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestS3.R
\name{plot.crestObj}
\alias{plot.crestObj}
\title{Plot the reconstructions.}
\usage{
\method{plot}{crestObj}(
  x,
  climate = x$parameters$climate,
  uncertainties = x$parameters$uncertainties,
  optima = TRUE,
  add_modern = FALSE,
  xlim = NA,
  ylim = NA,
  save = FALSE,
  loc = getwd(),
  ...
)
}
\arguments{
\item{x}{A crestObj produced by the crest.reconstruct() or crest() functions.}

\item{climate}{The climate variables to plot (default is all the reconstructed variables from x)}

\item{uncertainties}{A (vector of) threshold value(s) indicating the error
bars that should be calculated (default are the values stored in x).}

\item{optima}{A boolean to indicate whether to plot the optimum (TRUE) or the
mean (FALSE) estimates.}

\item{add_modern}{Adds the modern climate values to the plot.}

\item{xlim}{the x limits (x1, x2) of the plot.  Note that \code{x1 > x2}
    is allowed and leads to a \sQuote{reversed axis}.

    The default value, \code{NULL}, indicates that the range of the
    \link{finite} values to be plotted should be used.}

\item{ylim}{the y limits of the plot.}

\item{save}{A boolean to indicate if the diagram shoud be saved as a pdf file.
Default is FALSE.}

\item{loc}{An absolute or relative path that indicates the folder where the
diagram(s) hould be saved. Also used to specify the name of the file.
Default: the file is saved in the working directory with a file
created for each variable as variable.pdf.}

\item{...}{other \link[graphics]{graphical parameters} (see \code{\link[graphics]{par}} and
    section \sQuote{Details} below).}
}
\description{
Plot the reconstructions and their uncertainties
}
