% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_UsefulFunctions.R
\name{plot_map_eqearth}
\alias{plot_map_eqearth}
\title{Plots raster data in equal earth projection.}
\usage{
plot_map_eqearth(
  dat,
  ext = raster::extent(dat),
  zlim = range(raster::values(dat), na.rm = TRUE),
  col = viridis::viridis(20),
  brks.pos,
  brks.lab = brks.pos,
  npoints = 15,
  nlines = 9,
  title = ""
)
}
\arguments{
\item{dat}{The raster data to plot.}

\item{ext}{The extent to use to plot the data. (default is extent of dat)}

\item{zlim}{The range of the values to plot. (default is estimated from dat)}

\item{col}{The color gradient to use. (default is viridis)}

\item{brks.pos}{The position where to draw tick marks on the legend}

\item{brks.lab}{The labels to add where the tickmarks are draw (default is
tickmarks position)}

\item{npoints}{The number of points used to draw the polygons and lines along
each dimension. (default is 15 for a smooth result)}

\item{nlines}{The number of coordinate lines to add in the background
'(default is 9)}

\item{title}{A description title (default is empty)}
}
\description{
Plots raster data in equal earth projection.
}
