% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reconstr}
\alias{reconstr}
\title{A \code{\link{crestObj}} ran with the example dataset.}
\format{
An object of class \code{crestObj} of length 5.
}
\usage{
reconstr
}
\description{
A \code{\link{crestObj}} ran with the example dataset. Useful to illustrate
many functions of the package.
}
\keyword{datasets}
